Grouping IT tests per API surface
