+++ examples/pom.xml
@@ -22,7 +22,7 @@
     <maven.compiler.source>${java.version}</maven.compiler.source>
     <maven.compiler.target>${java.version}</maven.compiler.target>
     <spotbugs.fail>false</spotbugs.fail>
-    <opentelemetry.version>0.14.0</opentelemetry.version>
+    <opentelemetry.version>1.14.0</opentelemetry.version>
     <opentelemetry-sdk-metrics.version>1.41.0</opentelemetry-sdk-metrics.version>
     <zipkin.version>2.16.3</zipkin.version>
   </properties>
