+++ sdk-tests/src/test/java/io/dapr/it/testcontainers/actors/TestActorsApplication.java
@@ -11,16 +11,24 @@
 limitations under the License.
 */
 
-package io.dapr.it.testcontainers;
+package io.dapr.it.testcontainers.actors;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
-public class TestWorkflowsApplication {
+@RestController
+public class TestActorsApplication {
 
   public static void main(String[] args) {
-    SpringApplication.run(TestWorkflowsApplication.class, args);
+    SpringApplication.run(TestActorsApplication.class, args);
   }
 
+  //Mocking the actuator health endpoint for the sidecar health check
+  @GetMapping("/actuator/health")
+  public String health(){
+    return "OK";
+  }
 }
